JavaDoc comments. Removed indexed add functions.
Added a lot of JavaDoc comments. Also, removed all functions which included an index in its add function. Now, the internal Lists will handle the adding, since the order of children doesn't affect the functionality of our program. Indexing is still sometimes important for retrieving, so that stays for now.